+++ apps/client/src/components/pages/parkingLot/DetailMenuButtons.tsx
@@ -3,15 +3,9 @@ import MenuIconListItem from '@/components/common/MenuIconListItem';
 import { PaddedSection } from '@repo/ui/components/common/CommonLayouts';
 import { MessageCircle, Star, ThumbsDown, ThumbsUp } from 'lucide-react';
 import ReservationButton from './ReservationButton';
-import {
-  AddFavoriteAction,
-  DeleteFavoriteAction,
-} from '@/actions/user/user-service';
-import { checkIsFavorite } from '@/actions/user/user-service';
-import { useAlertWithLoading } from '@/hooks/useAlertWithLoading';
 import AlertModal from '@repo/ui/components/common/AlertModal';
-import { useEffect, useState } from 'react';
 import DotSpinner from '@repo/ui/components/icon/DotSpinner';
+import { useFavoriteStatus } from '@/hooks/useFavoriteStatus';
 
 export default function DetailMenuButtons({
   hostUuid,
@@ -30,52 +24,22 @@ export default function DetailMenuButtons({
   baseFee: number;
   baseIntervalMinutes: number;
 }) {
-  const { alertModalOpen, setAlertModalOpen, modalMessage, handleAlert } =
-    useAlertWithLoading();
+  const {
+    isFavorited,
+    isInitialLoadComplete,
+    handleAddFavorite,
+    alertModalOpen,
+    setAlertModalOpen,
+    modalMessage,
+  } = useFavoriteStatus({ parkingLotUuid });
 
-  const [isFavorited, setIsFavorited] = useState<boolean>(false);
-  const [isInitialLoadComplete, setIsInitialLoadComplete] =
-    useState<boolean>(false);
-
-  useEffect(() => {
-    const fetchFavoriteStatus = async () => {
-      try {
-        const res = await checkIsFavorite(parkingLotUuid);
-        if (!res.success || res.data === undefined) {
-          setIsFavorited(false);
-          return;
-        }
-        setIsFavorited(res.data);
-      } catch (e) {
-        console.error('초기 즐겨찾기 상태 불러오기 오류:', e);
-        setIsFavorited(false);
-      } finally {
-        setIsInitialLoadComplete(true);
-      }
-    };
-    fetchFavoriteStatus();
-  }, [parkingLotUuid]);
-
-  const handleAddFavorite = async () => {
-    try {
-      let res;
-      if (isFavorited) {
-        res = await DeleteFavoriteAction(parkingLotUuid);
-      } else {
-        res = await AddFavoriteAction(parkingLotUuid);
-      }
-
-      if (res.success) {
-        handleAlert(res.data);
-        setIsFavorited((prev) => !prev);
-      } else {
-        handleAlert(res.message);
-      }
-    } catch (e) {
-      console.error('즐겨찾기 액션 오류:', e);
-      handleAlert('즐겨찾기 처리 중 오류가 발생했습니다.');
-    }
-  };
+  if (!isInitialLoadComplete) {
+    return (
+      <PaddedSection className="py-5 space-y-5 bg-white flex justify-center items-center h-[152px]">
+        <DotSpinner />
+      </PaddedSection>
+    );
+  }
 
   if (!isInitialLoadComplete) {
     return (
